Add js to multiplayer dialog
