+++ index.html
@@ -23,7 +23,7 @@
           <p>This will be some content!</p>
           <div class="action">
             <button id="single-player">Single Player</button>
-            <button>Multiplayer</button>
+            <button id="multiplayer">Multiplayer</button>
           </div>
           <div class="content__level-2">
             <div class="content__level-2--single-player" id="dialog-single-player">
@@ -68,7 +68,7 @@
                 </fieldset>
               </div>
             </div>
-            <div class="content__level-2--multi-player">
+            <div class="content__level-2--multi-player" id="dialog-multiplayer">
 
             </div>
           </div>
